streamlit_app/requirements.txt
streamlit==1.23.1
pandas
pymongo
pinecone-client
python-dotenv
sentence-transformers
transformers
torch
matplotlib
plotly
wordcloud
scikit-learn